Move UI components from appcore into main app
