+++ app/src/main/java/org/torproject/android/ui/components/NoPersonalizedLearningEditText.kt
@@ -1,4 +1,4 @@
-package org.torproject.android.core.ui
+package org.torproject.android.ui.components
 
 import android.content.Context
 import android.os.Build
