+++ app/src/main/java/org/torproject/android/ui/more/SettingsActivity.kt
@@ -3,7 +3,7 @@ package org.torproject.android.ui.more
 import android.os.Bundle
 import android.view.MenuItem
 import org.torproject.android.R
-import org.torproject.android.core.ui.BaseActivity
+import org.torproject.android.ui.components.BaseActivity
 
 class SettingsActivity : BaseActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
