+++ app/src/main/java/org/torproject/android/ui/more/SettingsPreferenceFragment.kt
@@ -16,7 +16,7 @@ import androidx.preference.PreferenceCategory
 import androidx.preference.PreferenceFragmentCompat
 import org.torproject.android.R
 import org.torproject.android.core.Languages
-import org.torproject.android.core.ui.BaseActivity
+import org.torproject.android.ui.components.BaseActivity
 import org.torproject.android.service.util.Prefs
 import org.torproject.android.ui.more.camo.CamoFragment
