+++ app/src/main/java/org/torproject/android/ui/v3onionservice/OnionServiceActivity.java
@@ -22,7 +22,7 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import org.torproject.android.R;
 import org.torproject.android.core.DiskUtils;
-import org.torproject.android.core.ui.BaseActivity;
+import org.torproject.android.ui.components.BaseActivity;
 import org.torproject.android.service.db.OnionServiceColumns;
 
 public class OnionServiceActivity extends BaseActivity {
